requirements.txt
duckdb
python-dotenv
fastapi
uvicorn 